[Feat/#111] MedicinesManager 생성
